+++ test/apiserver/module_tests/test_retrieval_of_umap_plots.sh
@@ -8,4 +8,13 @@ then
     curl "$query"
     exit 1
 fi
-rm _rows.json
+
+size=$(stat -c "%s" _rows.json)
+if [ $size -lt 1000000 ];
+then
+    echo "JSON file ($size bytes) is less than 1MB in size, as expected."
+    rm _rows.json
+else
+    echo "JSON file is unexpectedly large: $size bytes."
+    exit 1
+fi
